add option to sway to choose hdmi sink from monitors
In pipewire, it aggregates all sinks connected to the GPU to a single sink. Where the sink itself has multiple profiles for each output it can use. This change allows one to choose the profile (hardcoded for now) and the automatically sets the hdmi as sink since the name changes after choosing a new profile.